pyproject.toml
[project]
name = "lattice"
version = "0.1.0"
description = "Codebase intelligence engine and multi-session orchestrator for Claude Code"
authors = [
    { name = "Lattice Contributors" }
]
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.10",
    "pydantic-settings[yaml]>=2.0",
    "structlog>=25.0",
    "pyyaml>=6.0",
    "langgraph>=1.0",
    "langgraph-checkpoint-duckdb>=2.0",
    "faiss-cpu>=1.7",
    "langchain-anthropic>=1.0",
    "langchain-openai>=1.0",
    "langchain-ollama>=1.0",
    "networkx>=3.6.1",
    "click>=8.3.1",
    "rich>=13.0",
    "python-frontmatter>=1.1.0",
    "fastapi>=0.100",
    "uvicorn>=0.20",
    "faster-whisper>=1.0.0",
    "sounddevice>=0.5.0",
    "deepgram-sdk>=6.0.0",
    "scipy>=1.11",
    "pynput>=1.8.0",
    "tavily-python>=0.7",
    "PyGithub>=2.8",
    "mattermostautodriver>=11.5",
    "mcp>=1.26",
]

[project.scripts]
lattice = "lattice.cli.commands:cli"

[tool.pytest.ini_options]
addopts = "--import-mode=importlib -q"
testpaths = ["tests"]
asyncio_mode = "auto"
norecursedirs = ["fixtures"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "httpx>=0.24",
]
